requirements.txt
flask
langchain
langchain-text-splitters
langchain-community
faiss-cpu
PyPDF2
python-dotenv
sentence-transformers
google-generativeai